Move hostname customizer
