+++ docker/src/main/java/clocker/docker/entity/util/JcloudsHostnameCustomizer.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package clocker.docker.entity;
+package clocker.docker.entity.util;
 
 import java.util.List;
